# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Golang deployment in vercel serverless environment",
        "title": "Golang Vercel Deployment",
        "contact": {},
        "version": "1.0"
    },
    "host": "golang-vercel.vercel.app",
    "paths": {
...
        "/hello/:name": {
            "get": {
                "description": "Endpoint to Welcome user and say Hello \"Name\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Welcome"
                ],
                "summary": "Hello User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name in the URL param",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Request Error or parameter missing",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "When user not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    }
}